refactor(version): move version constants to separate file
