+++ packages/framework/src/constants/version.constants.ts
@@ -1,0 +1,4 @@
+import { version } from '../../package.json';
+
+export const SDK_VERSION = version;
+export const FRAMEWORK_VERSION = '2024-06-26';
