--- packages/framework/src/version.ts
@@ -1,4 +1,0 @@
-import packageJson from '../package.json';
-
-export const SDK_VERSION = packageJson.version;
-export const FRAMEWORK_VERSION = '2024-06-26';
